small fixes, batch calls
