+++ src/adapters/tholgar/ethereum/balance.ts
@@ -1,4 +1,6 @@
 import type { Balance, BalancesContext, BaseBalance, BaseContract, Contract } from '@lib/adapter'
+import { mapSuccessFilter } from '@lib/array'
+import { call } from '@lib/call'
 import { abi as erc20Abi } from '@lib/erc20'
 import { multicall } from '@lib/multicall'
 import type { Token } from '@lib/token'
@@ -45,14 +47,6 @@ const aura: Token = {
   symbol: 'AURA',
 }
 
-const warAuraLocker: Contract = {
-  chain: 'ethereum',
-  address: '0x7B90e043aaC79AdeA0Dbb0690E3c832757207a3B',
-}
-const warCvxLocker: Contract = {
-  chain: 'ethereum',
-  address: '0x700d6d24A55512c6AEC08820B49da4e4193105B3',
-}
 const warRedeemer: Contract = {
   chain: 'ethereum',
   address: '0x4787Ef084c1d57ED87D58a716d991F8A9CD3828C',
@@ -64,88 +58,58 @@ const stkWAR: Contract = {
   symbol: 'stkWAR',
 }
 
-export async function getWarlordVault(ctx: BalancesContext, vault: Contract): Promise<Balance> {
-  const [tWarBalance, warBalance] = await multicall({
-    ctx,
-    abi: erc20Abi.balanceOf,
-    calls: [
-      {
-        target: vault.address,
-        params: [ctx.address],
-      },
-      {
-        target: stkWAR.address,
-        params: [vault.address],
-      },
-    ],
-  })
-  const [tWarTotalSupply, warTotalSupply] = await multicall({
-    ctx,
-    abi: erc20Abi.totalSupply,
-    calls: [vault.address, (vault!.underlyings![0] as BaseContract).address].map(
-      (target) =>
-        ({
-          target,
-        }) as const,
-    ),
-  })
-  const [cvxLocked, auraLocked] = await multicall({
-    ctx,
-    abi: abi.getCurrentLockedTokens,
-    calls: [warCvxLocker.address, warAuraLocker.address].map(
-      (target) =>
-        ({
-          target,
-        }) as const,
-    ),
-  })
-  const [auraQueued, cvxQueued] = await multicall({
-    ctx,
-    abi: abi.queuedForWithdrawal,
-    calls: [aura.address, cvx.address].map(
-      (token) =>
-        ({
-          target: warRedeemer.address,
-          params: [token],
-        }) as const,
-    ),
-  })
+export async function getWarlordVault(ctx: BalancesContext, vault: Contract, lockers: Contract[]): Promise<Balance> {
+  const [tWarBalance, warBalance, tWarTotalSupply, warTotalSupply, tokensLocked, tokensQueued] = await Promise.all([
+    call({
+      ctx,
+      target: vault.address,
+      params: [ctx.address],
+      abi: erc20Abi.balanceOf,
+    }),
+    call({
+      ctx,
+      target: stkWAR.address,
+      params: [vault.address],
+      abi: erc20Abi.balanceOf,
+    }),
+    call({
+      ctx,
+      target: vault.address,
+      abi: erc20Abi.totalSupply,
+    }),
+    call({
+      ctx,
+      target: (vault!.underlyings![0] as BaseContract).address,
+      abi: erc20Abi.totalSupply,
+    }),
+    multicall({
+      ctx,
+      abi: abi.getCurrentLockedTokens,
+      calls: lockers.map((locker) => ({ target: locker.address })),
+    }),
+    multicall({
+      ctx,
+      calls: [aura, cvx].map((token) => ({ target: warRedeemer.address, params: [token.address] }) as const),
+      abi: abi.queuedForWithdrawal,
+    }),
+  ])
 
-  let cvxUserBalanceAmount: bigint = 0n
-  let auraUserBalanceAmount: bigint = 0n
-  if (warBalance.success && warTotalSupply.success && tWarBalance.success && tWarTotalSupply.success) {
-    if (cvxLocked.success && cvxQueued.success) {
-      cvxUserBalanceAmount =
-        ((((cvxLocked.output - cvxQueued.output) * warBalance.output) / warTotalSupply.output) * tWarBalance.output) /
-        tWarTotalSupply.output
-    }
-    if (auraLocked.success && auraQueued.success) {
-      auraUserBalanceAmount =
-        ((((auraLocked.output - auraQueued.output) * warBalance.output) / warTotalSupply.output) * tWarBalance.output) /
-        tWarTotalSupply.output
-    }
+  const [auraLocked, cvxLocked] = mapSuccessFilter(tokensLocked, (res) => (res.output > 0n ? res.output : 0n))
+  const [auraQueued, cvxQueued] = mapSuccessFilter(tokensQueued, (res) => (res.output > 0n ? res.output : 0n))
+
+  const auraUserBalance: BaseBalance = {
+    ...aura,
+    amount: ((((auraLocked - auraQueued) * warBalance) / warTotalSupply) * tWarBalance) / tWarTotalSupply,
   }
 
   const cvxUserBalance: BaseBalance = {
-    amount: cvxUserBalanceAmount,
-    token: cvx.address,
-    chain: ctx.chain,
-    address: cvx.address,
-    symbol: cvx.symbol,
-    decimals: cvx.decimals,
-  }
-  const auraUserBalance: BaseBalance = {
-    amount: auraUserBalanceAmount,
-    token: aura.address,
-    chain: ctx.chain,
-    address: aura.address,
-    symbol: aura.symbol,
-    decimals: aura.decimals,
+    ...cvx,
+    amount: ((((cvxLocked - cvxQueued) * warBalance) / warTotalSupply) * tWarBalance) / tWarTotalSupply,
   }
 
   return {
     ...vault,
-    amount: tWarBalance.success ? tWarBalance.output : 0n,
+    amount: tWarBalance,
     underlyings: [cvxUserBalance, auraUserBalance],
     category: 'stake',
     rewards: undefined,
